+++ src/ZEO/scripts/cache_simul.py
@@ -75,7 +75,7 @@ def main(args=None):
     # Read trace file, simulating cache behavior.
     f_read = f.read
     unpack = struct.unpack
-    FMT = ">iiH8s8s"
+    FMT = ">IiH8s8s"
     FMT_SIZE = struct.calcsize(FMT)
     assert FMT_SIZE == 26